requirements.txt
accelerate==1.6.0
aiohttp==3.11.18
aiohttp==3.11.13
apex==0.9.10dev
byted_wandb==0.13.87
cachetools==2.0.1
cachetools==5.5.2
clearml==2.0.2
cloudpickle==3.1.1
codetiming==1.4.0
cupy_cuda12x==13.4.1
datasets==3.5.0
distribute==0.6.26
einops==0.8.1
fastapi==0.116.1
filelock==3.18.0
flash_attn==2.7.4.post1
func_timeout==4.3.5
huggingface_hub==0.30.2
human_eval.egg==info
hydra-core==1.3.2
json_repair==0.42.0
latex2sympy2_extended==1.10.1
liger_kernel==0.5.8
mathruler==0.1.0
mlflow==3.1.4
modelscope==1.28.0
numpy==1.11.1
numpy==1.6.2
omegaconf==2.3.0
openai==1.97.1
packaging==25.0
pandas==0.14.1
peft==0.15.2
Pillow==2.5.1
Pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pyext==0.6
pylatexenc==2.10
qwen_vl_utils==0.0.11
Requests==2.32.4
safetensors==0.5.3
setuptools==28.8.0
setuptools==32.1.2
setuptools==75.8.0
starlette==0.47.2
swanlab==0.6.7
sympy==1.13.1
tensordict==0.6.2
torch==2.6.0+cu118
torch_npu==2.5.1
torchdata==0.11.0
tqdm==4.67.1
uvicorn==0.35.0
